refactor(mindweaver): move packages/pkg to packages/mindweaver/shared
Consolidates shared utilities into the mindweaver module structure.
This prepares for independent component versioning where future modules
(lsp, imex, web) will import from the mindweaver module.
