+++ packages/mindweaver/shared/config/README.md
