+++ packages/mindweaver/shared/config/config_test.go
