+++ packages/mindweaver/shared/dberrors/errors.go
