+++ packages/mindweaver/shared/sqlcext/bulk.go
