+++ packages/mindweaver/shared/sqlcext/cte.go
