+++ packages/mindweaver/shared/sqlcext/fts.go
