+++ packages/mindweaver/shared/sqlcext/fts_test.go
