+++ packages/mindweaver/shared/transport/http.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/nkapatos/mindweaver/pkg/types"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/shared/types"
 )
 
 // HTTPTransportConfig configures the HTTP transport
