+++ packages/mindweaver/shared/utils/etag.go
