+++ packages/mindweaver/shared/utils/null_helpers.go
